refactor: split route.ts into focused modules
Extract helper functions from the 1334-line route.ts into
four dedicated modules:

- types.ts      — Zod schemas, derived types, TMDB interface
- tmdb.ts       — TMDB fetch, cosine scoring, JIT seeding
- embedding.ts  — query enrichment + embedding creation
- recommendation.ts — OpenAI recommendation, AI descriptions, poster lookups

route.ts reduced to 419 lines containing only the POST and
GET handlers. All 45 tests pass, TypeScript and ESLint clean.
